typecast some return values
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@1542 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdePkg/Library/BaseTimerLibLocalApic/x86TimerLib.c
@@ -210,7 +210,7 @@ GetPerformanceCounter (
   VOID
   )
 {
-  return (UINT32)InternalX86GetTimerTick (InternalX86GetApicBase ());
+  return (UINT64)(UINT32)InternalX86GetTimerTick (InternalX86GetApicBase ());
 }
 
 /**
@@ -255,5 +255,5 @@ GetPerformanceCounterProperties (
     *EndValue = 0;
   }
 
-  return InternalX86GetTimerFrequency (ApicBase);
+  return (UINT64)InternalX86GetTimerFrequency (ApicBase);
 }
